fix: refactor plugin redux store
